requirements.txt
jmcomic
cbz
rich
requests
pyyaml
